requirements.txt
paho-mqtt
pymongo
requests
PyJWT
